Promote `files` as a top level command
