+++ test/spoom/cli/commands/config_test.rb
@@ -144,50 +144,6 @@ module Spoom
              * .ru
           MSG
         end
-
-        def test_display_files_from_config
-          use_sorbet_config(PROJECT, ".")
-          out, _ = run_cli(PROJECT, "config files")
-          assert_equal(<<~MSG, out)
-            Files matching `sorbet/config`:
-             * errors/errors.rb
-             * lib/defs.rb
-             * lib/hover.rb
-             * lib/refs.rb
-             * lib/sigs.rb
-             * lib/symbols.rb
-             * lib/types.rb
-          MSG
-        end
-
-        def test_display_files_from_config_with_ignored_files
-          use_sorbet_config(PROJECT, <<~CFG)
-            .
-            --ignore=efs
-            --ignore=errors
-          CFG
-          out, _ = run_cli(PROJECT, "config files")
-          assert_equal(<<~MSG, out)
-            Files matching `sorbet/config`:
-             * lib/hover.rb
-             * lib/sigs.rb
-             * lib/symbols.rb
-             * lib/types.rb
-          MSG
-        end
-
-        def test_display_files_from_config_with_allowed_exts
-          use_sorbet_config(PROJECT, <<~CFG)
-            .
-            --allowed-extension=.rake
-          CFG
-          out, _ = run_cli(PROJECT, "config files")
-          assert_equal(<<~MSG, out)
-            Files matching `sorbet/config`:
-             * task1.rake
-             * task2.rake
-          MSG
-        end
       end
     end
   end
