+++ test/support/project/task1.rake
@@ -1,0 +1,2 @@
+# typed: false
+# frozen_string_literal: true
