+++ test/support/project/task2.rake
@@ -1,0 +1,2 @@
+# typed: strong
+# frozen_string_literal: true
